backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx
pydantic
soundfile
torchaudio
websockets
python-multipart
